Persist event filter reducer in chrome extension
