+++ cfg/dist.js
@@ -12,7 +12,7 @@ const config = Object.assign({}, baseConfig, {
 config.entry.app.push(path.join(__dirname, '../src/index'))
 
 config.plugins.push(...[
-  new UglifyJSPlugin({ sourceMap: true }),
+  //new UglifyJSPlugin({ sourceMap: true }),
   new webpack.optimize.AggressiveMergingPlugin(),
   new webpack.NoEmitOnErrorsPlugin()
 ])
